Add custom BIOS implementation, separate emulation core from platform and UI code.

+++ include/platform/audio.h
@@ -4,6 +4,6 @@
 
 void audioInit();
 void audioCleanup();
-u16 audioGetSampleRate();
+u32 audioGetSampleRate();
 void audioClear();
 void audioPlay(u32* buffer, long samples);
