+++ source/platform/sdl/audio_sdl.cpp
@@ -39,7 +39,7 @@ void audioCleanup() {
     }
 }
 
-u16 audioGetSampleRate() {
+u32 audioGetSampleRate() {
     return 44100;
 }
